requirements.txt
streamlit
# google-cloud-aiplatform
google-generativeai
google-cloud-firestore
google-auth
pydantic
pillow
python-dotenv
fpdf